+++ bin/node/src/commands/mod.rs
@@ -187,8 +187,12 @@ pub struct NtxBuilderConfig {
     /// Maximum number of crashes before an account deactivated.
     ///
     /// Once this limit is reached, no new transactions will be created for this account.
-    #[arg(long = "ntx-builder.max-actor-crashes", default_value_t = 10, value_name = "NUM")]
-    pub max_actor_crashes: usize,
+    #[arg(
+        long = "ntx-builder.max-account-crashes",
+        default_value_t = 10,
+        value_name = "NUM"
+    )]
+    pub max_account_crashes: usize,
 
     /// Directory for the ntx-builder's persistent database.
     ///
@@ -221,7 +225,7 @@ impl NtxBuilderConfig {
         .with_tx_prover_url(self.tx_prover_url)
         .with_script_cache_size(self.script_cache_size)
         .with_idle_timeout(self.idle_timeout)
-        .with_max_actor_crashes(self.max_actor_crashes)
+        .with_max_account_crashes(self.max_account_crashes)
     }
 }